fix segfault on ways with no nodes (happend in last OHM dump), add DenseMemoryIndex which stores node location in a single continuous dense block of memory, change option --store-locations-on-disk to --store-locations={disk-sparse, disk-dense, mem-dense, mem-flex (default)}, reduce memory consumption of RelationHandler by only using 32bit way ids (can be easily changed to support 64 bit again by holding a second map with 64bit ids, and choosing them based on the stored / requested ID) and by storing node IDs as 32 bit integer: node IDs < 2^31 - 1 are directly stored unchanged, and all other node IDs are stored with the upper 31 in a first byte, marked by a 1 in the highest bit, and the remaining 32bit in the second byte. This means we only support 63bit node IDs, but I think we can safely assume that OSM will not reach node ids > 2^63 for some years to come :)

+++ include/osm2rdf/osm/RelationHandler.h
@@ -33,12 +33,12 @@ class RelationHandler : public osmium::handler::Handler {
   void setLocationHandler(osm2rdf::osm::LocationHandler* locationHandler);
   bool hasLocationHandler() const;
   osmium::Location get_node_location(const uint64_t nodeId) const;
-  std::vector<uint64_t> get_noderefs_of_way(const uint64_t wayId);
+  std::vector<uint64_t> get_noderefs_of_way(const uint32_t wayId);
 
  protected:
   osm2rdf::config::Config _config;
   osm2rdf::osm::LocationHandler* _locationHandler = nullptr;
-  std::unordered_map<uint64_t, std::vector<uint64_t>> _ways;
+  std::unordered_map<uint32_t, std::vector<uint32_t>> _ways;
   bool _firstPassDone = false;
 };
 }
